feat : 메인 페이지 라우팅 경로 home->main으로 다시 원상 복구
