+++ src/app/main/GenreColorArr.css
